updated server to use clusters, mongo to return lean for initial get

+++ server/index.js
@@ -1,43 +1,56 @@
 /* eslint-disable radix */
-const express = require('express');
-const path = require('path');
-const cors = require('cors');
-const bodyParser = require('body-parser');
+var cluster = require('cluster');
 
-const app = express();
-const mongo = require('./db/mongoSeedScripts/readMongo');
+// Code to run if we're in the master process
+if (cluster.isMaster) {
+    // Count the machine's CPUs
+    var cpuCount = require('os').cpus().length;
+    // Create a worker for each CPU
+    for (var i = 0; i < cpuCount; i += 1) {
+        cluster.fork();
+    }
+} else {
+  require('newrelic');
+  const express = require('express');
+  const path = require('path');
+  const cors = require('cors');
+  const bodyParser = require('body-parser');
 
-const port = 5001;
-app.use(cors());
-app.use(express.static(path.join(__dirname, '../client/public/')));
+  const app = express();
+  const mongo = require('./db/mongoSeedScripts/readMongo');
 
-app.get('/restaurants/:restaurantId/reviews', (req, res) => {
-  mongo.getReviews(res, req.params.restaurantId);
-});
+  const port = 5001;
+  app.use(cors());
+  app.use(express.static(path.join(__dirname, '../client/public/')));
 
-app.patch('/restaurants/:restaurantId/reviews/:reviewId', (req, res) => {
-  const voteInfo = {
-    restaurant_id: Number(req.params.restaurantId),
-    review_id: Number(req.params.reviewId),
-    voteType: req.query.value,
-    voteStatus: `${req.query.value.split('_')[0]}_count`,
-    voted: Number(req.query.voted),
-  };
-  mongo.updateVoteCount(res, voteInfo);
-});
+  app.get('/restaurants/:restaurantId/reviews', (req, res) => {
+    mongo.getReviews(res, req.params.restaurantId);
+  });
 
-const jsonParser = bodyParser.json();
+  app.patch('/restaurants/:restaurantId/reviews/:reviewId', (req, res) => {
+    const voteInfo = {
+      restaurant_id: Number(req.params.restaurantId),
+      review_id: Number(req.params.reviewId),
+      voteType: req.query.value,
+      voteStatus: `${req.query.value.split('_')[0]}_count`,
+      voted: Number(req.query.voted),
+    };
+    mongo.updateVoteCount(res, voteInfo);
+  });
 
-app.post('/restaurants/:name', jsonParser, (req, res) => {
-  db.postRestaurant(req, res);
-});
+  const jsonParser = bodyParser.json();
 
-app.put('/restaurants/:restaurantId', jsonParser, (req, res) => {
-  db.updateRestaurant(req, res);
-});
+  app.post('/restaurants/:name', jsonParser, (req, res) => {
+    db.postRestaurant(req, res);
+  });
 
-app.delete('/restaurants/:restaurantId', jsonParser, (req, res) => {
-  db.deleteRestaurant(req, res);
-});
+  app.put('/restaurants/:restaurantId', jsonParser, (req, res) => {
+    db.updateRestaurant(req, res);
+  });
 
-app.listen(port, () => console.log(`SQUAWK listening on port ${port}!`));
+  app.delete('/restaurants/:restaurantId', jsonParser, (req, res) => {
+    db.deleteRestaurant(req, res);
+  });
+
+  app.listen(port, () => console.log(`Worker ${cluster.worker.id} listening on port ${port}!`));
+}
